add schelued GamesTaskService and logging to it and GamesController
This adds a new service class with a "killOldGames" method that runs
every 10 seconds. Ultimately this will be responsible for clearing out
old games, but for now it just writes a message to the logs.

I also added some basic logging to the GamesController to easily
indicate requests being made.