only calculate open areas when needed
Previously, the server bot logic would calculate the area of each open
section of the board on EVERY move, even if it never used that data in
decididing its move. This was woefully inneficient.

This commit changes it to only calculate that information just in time,
when it's about to use it.